% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMM-ICs.R
\name{MMComparisonPlot}
\alias{MMComparisonPlot}
\title{Multiple Model Comparison Plot}
\usage{
MMComparisonPlot(
  ICweights,
  models1,
  models2,
  main = "Multiple Models Comparison",
  ylab = "Probability",
  xlab = "Participant",
  groupNames = c("models1", "models2"),
  inset = c(-0.5, -0.25),
  colours = c("darkgreen", "lightblue")
)
}
\arguments{
\item{ICweights}{A numeric data frame/matrix/array that contains the model comparison information criterion (e.g., AIC, BIC, LOOIC) of each participant (rows) and each model (column).}

\item{models1, models2}{A vector of column names or column numbers indicating a group of models to be compared together. Note to be cautious when comparing groups that have an uneven number of models.}

\item{main}{Title of plot.}

\item{ylab}{y axis label.}

\item{xlab}{x axis label.}

\item{groupNames}{character vector of length 2 describing the names of the two groups of models (to be shown on the plot legend).}

\item{inset}{determines the position of the plot legend.}

\item{colours}{character vector of length 2 containing valid r graphics colours.}
}
\value{
A stacked bar plot
}
\description{
Compare the relative probability of groups of models.
}
\examples{
MMComparisonPlot(weightedICs(exampleData), models1 = c("Model 1", "Model 2"), models2 = c("Model 3", "Model 4"))
MMComparisonPlot(weightedICs(exampleData), models1 = 1:2,models2 = 3:4)
}
