% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\name{testData}
\alias{testData}
\title{Example information criterion data}
\usage{
testData()
}
\description{
Example information criterion data that that shows you the format necessary to be used in the weightedIC() function.
}
