% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-ICs.R
\name{weightedICs}
\alias{weightedICs}
\title{Get the weighted probabilities of different models for each participant}
\usage{
weightedICs(ICs, bySubject = TRUE)
}
\arguments{
\item{ICs}{a numeric matrix/array/data frame containing the raw model comparison criterion values (e.g., AIC, BIC, LOOIC) for each model (columns) and participant (rows).}
}
\value{
A matrix of probabilities corresponding to each participant (rows) and model (columns). To plot this output, see ?plotWeightedICs() and ?compareMM-ICs().
}
\description{
Takes the raw model comparison information criterion values (e.g., AIC, BIC, LOOIC)
for each model (columns) and participant (rows) and then converts them into a relitive probability
of that model being the best performing model according to that model comparison criterion.
}
\examples{
weightedICs(exampleData)
}
