% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeightedICs.R
\name{plotWeightedICs}
\alias{plotWeightedICs}
\title{Plot weighted model comparisons}
\usage{
plotWeightedICs(
  ICweights,
  main = "Weighted Model Comparison",
  ylab = "Probability",
  xlab = "Participant",
  colours = c("#800000", "#0000FF", "#800080", "#FFC0CB", "#00FF00", "#00FFFF",
    "#FF00FF", "#FFFF00", "#FFA500", "#008080", "#FF0000", "#00FF00", "#4B0082",
    "#000080", "#808000", "#00FFFF", "#C0C0C0", "#808080", "#A52A2A", "#FFD700"),
  cex = 0.5,
  position = "bottom",
  seed = FALSE,
  inset = c(-0.5, -0.1)
)
}
\arguments{
\item{ICweights}{A numeric data frame/matrix/array that contains the model probabilities of each participant (rows) and each model (column).}

\item{main}{Title of plot.}

\item{ylab}{y axis label.}

\item{xlab}{x axis label.}

\item{colours, seed}{Character vector of base r colours. Default generates colours for each variable randomly,  You can also try different numeric values for "seed" to try out different random colour combinations.}

\item{cex}{Controls the size of the plot legend.}

\item{inset}{Determines the position of the plot legend.}

\item{groupNames}{Character vector of length 2 describing the names of the two groups of models (to be shown on the plot legend).}
}
\value{
A stacked bar plot.
}
\description{
Takes a matrix/array/data frame that contains model probabilities
for each model (columns) and participant (rows) (which can be created made using the "weightedICs" function) and turns it into a stacked bar
plot.
}
\examples{
plotWeightedICs(weightedICs(exampleData))
plotWeightedICs(weightedICs(exampleData), colours = (c("darkblue","darkgreen","lightgreen","lightblue")))
plotWeightedICs(weightedICs(exampleData), cex = 1, colours = (c("darkblue","darkgreen","lightgreen","lightblue")))

}
